requirements.txt
# TimeTrace - Application Usage Tracker
# Required Python packages

# GUI Framework
customtkinter>=5.2.0

# System monitoring
psutil>=5.9.0

# System tray support
pystray>=0.19.0
Pillow>=10.0.0

# Standard library modules (included with Python 3.10+):
# - sqlite3
# - json
# - threading
# - datetime
# - typing
